changed

+++ 0x05-pointers_arrays_strings/main.h
@@ -9,7 +9,7 @@
 int _putchar(char c);
 void reset_to_98(int *n);
 void swap_int(int *a, int *b); 
-size_t  _strlen(char *s);
+int _strlen(char *s);
